# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SimulationCategory =
  (typeof SimulationCategory)[keyof typeof SimulationCategory];

export const SimulationCategory = {
  engineering: "engineering",
  business: "business",
  healthcare: "healthcare",
  trades: "trades",
  technology: "technology",
  law: "law",
  science: "science",
  "life-advice": "life-advice",
  "creative-design": "creative-design",
} as const;

export type EnvironmentKind =
  (typeof EnvironmentKind)[keyof typeof EnvironmentKind];

export const EnvironmentKind = {
  "mech-lab": "mech-lab",
  "construction-site": "construction-site",
  "circuit-board": "circuit-board",
  "chem-plant": "chem-plant",
  "factory-floor": "factory-floor",
  "office-desk": "office-desk",
  "hospital-room": "hospital-room",
  "server-room": "server-room",
  "data-viz": "data-viz",
  courtroom: "courtroom",
  kitchen: "kitchen",
  "rooftop-solar": "rooftop-solar",
  "wind-turbine": "wind-turbine",
  spa: "spa",
  "brain-lab": "brain-lab",
  observatory: "observatory",
  "bio-lab": "bio-lab",
  "ocean-research": "ocean-research",
} as const;

export interface Simulation {
  slug: string;
  title: string;
  category: SimulationCategory;
  tagline: string;
  description: string;
  topics: string[];
  environment: EnvironmentKind;
  available: boolean;
}

export interface MissionChoice {
  id: string;
  text: string;
  /** @nullable */
  rationale?: string | null;
}

/**
 * Free-form parameters specific to the environment (signage, props, highlights, numeric values, etc).
 */
export type MissionSceneParams = { [key: string]: unknown };

/**
 * Parameters that drive the 3D environment rendered for this mission.
 */
export interface MissionScene {
  environment: EnvironmentKind;
  /** Free-form parameters specific to the environment (signage, props, highlights, numeric values, etc). */
  params: MissionSceneParams;
}

export interface Mission {
  id: string;
  simulationSlug: string;
  category: SimulationCategory;
  topic: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  difficulty: number;
  title: string;
  /** One-line setup that names the role the student is playing. */
  roleIntro: string;
  scenario: string;
  problem: string;
  constraints: string[];
  question: string;
  choices: MissionChoice[];
  scene: MissionScene;
  createdAt: string;
}

export interface GenerateMissionBody {
  learnerId: string;
  simulationSlug: string;
  /**
   * @minimum 1
   * @maximum 10
   * @nullable
   */
  targetDifficulty?: number | null;
  avoidTopics?: string[];
}

export interface SubmitMissionInput {
  learnerId: string;
  choiceId: string;
}

export interface MissionResult {
  missionId: string;
  correct: boolean;
  correctChoiceId: string;
  explanation: string;
  encouragement: string;
  xpAwarded: number;
  newStreak: number;
  newLevel: number;
  newTotalXp: number;
}

export interface HintResponse {
  hint: string;
}

export interface CategoryProgress {
  category: SimulationCategory;
  attempted: number;
  correct: number;
}

export interface Progress {
  learnerId: string;
  totalXp: number;
  level: number;
  streak: number;
  bestStreak: number;
  missionsCompleted: number;
  missionsCorrect: number;
  accuracy: number;
  topicsMastered: string[];
  nextLevelXp: number;
  categoryProgress: CategoryProgress[];
}

export interface MissionAttempt {
  id: string;
  missionId: string;
  title: string;
  topic: string;
  simulationSlug: string;
  category: SimulationCategory;
  difficulty: number;
  correct: boolean;
  xpAwarded: number;
  attemptedAt: string;
}

export type GetProgressParams = {
  /**
   * Anonymous learner identifier stored client-side
   */
  learnerId: string;
};

export type GetRecentMissionsParams = {
  learnerId: string;
};