# lib/api-zod/src/generated/types/categoryProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationCategory } from "./simulationCategory";

export interface CategoryProgress {
  category: SimulationCategory;
  attempted: number;
  correct: number;
}