# lib/api-zod/src/generated/types/missionAttempt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationCategory } from "./simulationCategory";

export interface MissionAttempt {
  id: string;
  missionId: string;
  title: string;
  topic: string;
  simulationSlug: string;
  category: SimulationCategory;
  difficulty: number;
  correct: boolean;
  xpAwarded: number;
  attemptedAt: string;
}